Conflicts:
	fannie/admin/labels/scanpage.php
	fannie/batches/CAP/CoopDealsUploadPage.php
	fannie/batches/UNFI/load-classes/UnfiUploadPage.php
	fannie/classlib2.0/FannieReportPage.php
	fannie/classlib2.0/FannieUploadPage.php
	fannie/classlib2.0/data/controllers/MeminfoController.php
	fannie/classlib2.0/data/controllers/ProdUpdateController.php
	fannie/classlib2.0/data/controllers/ProductsController.php
	fannie/classlib2.0/lib/FormLib.php
	fannie/item/ItemFlags.php
	fannie/item/mapping/index.php
	fannie/mem/modules/ContactPref.php
	fannie/mem/modules/MemCard.php
	fannie/reports/BatchReport/BatchReport.php
	fannie/reports/CustomerCount/CustomerCountReport.php
	fannie/reports/DepartmentMovement/DepartmentMovementReport.php
	fannie/reports/ManufacturerMovement/ManufacturerMovementReport.php
	fannie/reports/NonMovement/NonMovementReport.php
	fannie/reports/ProductMovement/ProductMovementModular.php
	fannie/reports/VendorMovement/VendorMovementReport.php
	fannie/src/defaultmenu.php
	fannie/src/navbar.dynamic.php
	pos/is4c-nf/install/plugins.php
	pos/is4c-nf/lib/LocalStorage/MemcacheStorage.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFilter/DefaultReceiptFilter.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFormat/DefaultReceiptFormat.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFormat/ItemFormat.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFormat/OtherFormat.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFormat/TenderFormat.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptFormat/TotalFormat.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptSort/DefaultReceiptSort.php
	pos/is4c-nf/lib/ReceiptBuilding/ReceiptTag/DefaultReceiptTag.php
	pos/is4c-nf/lib/ReceiptBuilding/TenderReports/TenderReport.php
	pos/is4c-nf/lib/ReceiptBuilding/TenderReports/WfcTenderReport.php
	pos/is4c-nf/lib/Search/Products/DefaultProductSearch.php
	pos/is4c-nf/lib/Search/Products/ProductUserSearch.php
	pos/is4c-nf/plugins/AllItemSearch/AllItemProductSearch.php
	pos/is4c-nf/plugins/Paycards/Paycards.php
	pos/is4c-nf/plugins/PriceCheck/PriceCheckPage.php
	pos/is4c-nf/plugins/VirtualCoupon/VirtCoupDisplay.php

+++ fannie/cron/Suspensions/reactivate.ar.php
@@ -31,7 +31,20 @@ include($FANNIE_ROOT.'src/SQLManager.php');
 
 /* HELP
 
-   This script activates members with equity paid in full
+   This script activates members with store-charge account (ar)
+   up-to-date, i.e.
+   ar_live_balance.balance < AR_EOM_Summary.twoMonthBalance
+
+   When/how-often can/should it be run? Daily?
+
+*/
+
+/* --COMMENTS - - - - - - - - - - - - - - - - - - - - - - - - - -
+ *
+ * 18Oct12 EL Keep this comment block from appearing in the Help popup.
+ *             Reformat first SQL statement.
+ * 17Jun12 EL Fix Help to make it appropriate to this program.
+ *             Was a copy of reactivate.equity.php.
 */
 
 set_time_limit(0);
@@ -44,18 +57,18 @@ $TRANS = $FANNIE_TRANS_DB . ($FANNIE_SERVER_DBMS=="MSSQL" ? 'dbo.' : '.');
 
 $meminfoQ = "UPDATE meminfo AS m LEFT JOIN
 	    custdata AS c ON m.card_no=c.CardNo
-	    LEFT JOIN {$TRANS}newBalanceToday_cust AS s
-	    ON c.cardno=s.memnum LEFT JOIN suspensions AS p
-	    ON c.cardno=p.cardno LEFT JOIN {$TRANS}AR_EOM_Summary_cache AS a
+	    LEFT JOIN {$TRANS}ar_live_balance AS s
+	    ON c.cardno=s.card_no LEFT JOIN suspensions AS p
+	    ON c.cardno=p.cardno LEFT JOIN {$TRANS}AR_EOM_Summary AS a
 	    ON m.card_no=a.cardno
 	    SET m.ads_OK=p.mailflag
 	    WHERE c.Type = 'INACT' and p.reasoncode IN (1)
-	    AND s.balance < a.twoMonthBalance";
+		AND s.balance < a.twoMonthBalance";
 $sql->query($meminfoQ);
 
-$custQ = "UPDATE custdata AS c LEFT JOIN {$TRANS}newBalanceToday_cust AS s
-	    ON c.CardNo=s.memnum LEFT JOIN suspensions AS p
-	    ON c.CardNo=p.cardno LEFT JOIN {$TRANS}AR_EOM_Summary_cache AS a
+$custQ = "UPDATE custdata AS c LEFT JOIN {$TRANS}ar_live_balance AS s
+	    ON c.CardNo=s.card_no LEFT JOIN suspensions AS p
+	    ON c.CardNo=p.cardno LEFT JOIN {$TRANS}AR_EOM_Summary AS a
 	    ON c.CardNo=a.cardno
 	    SET c.Discount=p.discount,c.memDiscountLimit=p.chargelimit,
 	    c.memType=p.memtype1,c.Type=p.memtype2,chargeOk=1
@@ -65,7 +78,7 @@ $custQ = "UPDATE custdata AS c LEFT JOIN {$TRANS}newBalanceToday_cust AS s
 $sql->query($custQ);
 
 $histQ = "insert into suspension_history
-	    select 'automatic',".$sql->now().",
+	    select 'AR paid',".$sql->now().",
 	    'Account reactivated',c.CardNo,0 from
 	    suspensions as s left join
 	    custdata as c on s.cardno=c.CardNo
@@ -85,7 +98,9 @@ while($clearW = $sql->fetch_row($clearR)){
 }
 $cns = rtrim($cns,",").")";
 
-$delQ = "DELETE FROM suspensions WHERE cardno IN $cns";
-$delR = $sql->query($delQ);
+if (strlen($cns) > 2){
+	$delQ = "DELETE FROM suspensions WHERE cardno IN $cns";
+	$delR = $sql->query($delQ);
+}
 
 ?>
